+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 4.4.0 - 2022-08-xx
 ### Fixed
 - Usernames shown as IDs instead of readable names #286
+- Faster report creation #293
 
 ## 4.3.1 - 2022-06-25
 ### Fixed