+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Usernames shown as IDs instead of readable names #286
 - Faster report creation #293
 - Notification issue for occ dataload #294
+- Share Link doesn't include port number #298
 
 ## 4.3.1 - 2022-06-25
 ### Fixed